+++ src/connectedmachine/azext_connectedmachine/tests/latest/test_RunCommand_scenario.py
@@ -41,9 +41,9 @@ class RunCommandScenarioTest(ScenarioTest):
     def test_run_command(self):
         rand_string = 'test'
         self.kwargs.update({
-            'machine': 'testmachine2',
-            'rg': 'ytongtest2',
-            'location': 'eastus2euap',
+            'machine': 'testmachine',
+            'rg': 'ytongtest',
+            'location': 'eastus',
             'subscription': '00000000-0000-0000-0000-000000000000',
             'runcommand': 'myRunCommand',
         })