+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/pool/_create.py
@@ -25,12 +25,15 @@ class Create(AAZCommand):
 
     :example: Create a pool with stop on disconnect enabled
         az devcenter admin pool create --location "eastus" --devbox-definition-name "WebDevBox" --network-connection-name "Network1-westus2" --pool-name "DevPool" --project-name "DevProject" --resource-group "rg1" --local-administrator "Enabled" --virtual-network-type "Unmanaged" --single-sign-on-status "Enabled" --stop-on-disconnect status="Enabled" grace-period-minutes="60"
+
+    :example: Create pool using image definition
+        az devcenter admin pool create --location "eastus" --network-connection-name "Network1-westus2" --pool-name "DevPool" --project-name "DevProject" --resource-group "rg1" --local-administrator "Enabled" --virtual-network-type "Unmanaged" --single-sign-on-status "Enabled" --devbox-definition-type "Value" --devbox-definition-image-reference id="/subscriptions/0ac520ee-14c0-480f-b6c9-0a90c58ffff/resourceGroups/rg1/providers/Microsoft.DevCenter/projects/DevProject/images/~Catalog~image2~project-sample-2" --devbox-definition-sku name=""general_i_8c32gb256ssd_v2"
     """
 
     _aaz_info = {
-        "version": "2024-05-01-preview",
+        "version": "2024-10-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/pools/{}", "2024-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/pools/{}", "2024-10-01-preview"],
         ]
     }
 
@@ -95,13 +98,65 @@ class Create(AAZCommand):
         tags = cls._args_schema.tags
         tags.Element = AAZStrArg()
 
+        # define Arg Group "DevBoxDefinition"
+
+        _args_schema = cls._args_schema
+        _args_schema.devbox_definition_image_reference = AAZObjectArg(
+            options=["-i", "--devbox-definition-image-reference"],
+            arg_group="DevBoxDefinition",
+            help="Image reference information for a definition of the machines that are created from this pool. Will be ignored if the parameter devbox-definition-type is Reference or not provided.",
+            is_preview=True,
+        )
+        _args_schema.devbox_definition_sku = AAZObjectArg(
+            options=["-s", "--devbox-definition-sku"],
+            arg_group="DevBoxDefinition",
+            help="The SKU for Dev Boxes created from the Pool. Will be ignored if the parameter devbox-definition-type is Reference or not provided.",
+            is_preview=True,
+        )
+
+        devbox_definition_image_reference = cls._args_schema.devbox_definition_image_reference
+        devbox_definition_image_reference.id = AAZStrArg(
+            options=["id"],
+            help="Image ID, or Image version ID. When Image ID is provided, its latest version will be used.",
+        )
+
+        devbox_definition_sku = cls._args_schema.devbox_definition_sku
+        devbox_definition_sku.capacity = AAZIntArg(
+            options=["capacity"],
+            help="If the SKU supports scale out/in then the capacity integer should be included. If scale out/in is not possible for the resource this may be omitted.",
+        )
+        devbox_definition_sku.family = AAZStrArg(
+            options=["family"],
+            help="If the service has different generations of hardware, for the same SKU, then that can be captured here.",
+        )
+        devbox_definition_sku.name = AAZStrArg(
+            options=["name"],
+            help="The name of the SKU. E.g. P3. It is typically a letter+number code",
+            required=True,
+        )
+        devbox_definition_sku.size = AAZStrArg(
+            options=["size"],
+            help="The SKU size. When the name field is the combination of tier and some other value, this would be the standalone code. ",
+        )
+        devbox_definition_sku.tier = AAZStrArg(
+            options=["tier"],
+            help="This field is required to be implemented by the Resource Provider if the service has more than one tier, but is not required on a PUT.",
+            enum={"Basic": "Basic", "Free": "Free", "Premium": "Premium", "Standard": "Standard"},
+        )
+
         # define Arg Group "Properties"
 
         _args_schema = cls._args_schema
         _args_schema.devbox_definition_name = AAZStrArg(
             options=["-d", "--devbox-definition-name"],
             arg_group="Properties",
-            help="Name of a dev box definition in parent project of this pool.",
+            help="Name of a dev box definition in parent project of this pool. Will be ignored if the parameter devbox-definition-type is Value.",
+        )
+        _args_schema.devbox_definition_type = AAZStrArg(
+            options=["-t", "--devbox-definition-type"],
+            arg_group="Properties",
+            help="Indicates if the pool is created from an existing Dev Box Definition or if one is provided directly.",
+            enum={"Reference": "Reference", "Value": "Value"},
         )
         _args_schema.display_name = AAZStrArg(
             options=["--display-name"],
@@ -122,7 +177,7 @@ class Create(AAZCommand):
             enum={"Disabled": "Disabled", "Enabled": "Enabled"},
         )
         _args_schema.managed_virtual_network_regions = AAZListArg(
-            options=["--managed-virtual-network-regions", "-m"],
+            options=["-m", "--managed-virtual-network-regions"],
             arg_group="Properties",
             help="The regions of the managed virtual network (required when managedNetworkType is Managed).",
         )
@@ -141,6 +196,11 @@ class Create(AAZCommand):
             options=["--stop-on-disconnect"],
             arg_group="Properties",
             help="Stop on disconnect configuration settings for dev boxes created in this pool.",
+        )
+        _args_schema.stop_on_no_connect = AAZObjectArg(
+            options=["--stop-on-no-connect"],
+            arg_group="Properties",
+            help="Stop on no connect configuration settings for Dev Boxes created in this pool.",
             is_preview=True,
         )
         _args_schema.virtual_network_type = AAZStrArg(
@@ -161,6 +221,17 @@ class Create(AAZCommand):
         stop_on_disconnect.status = AAZStrArg(
             options=["status"],
             help="Whether the feature to stop the dev box on disconnect once the grace period has lapsed is enabled.",
+            enum={"Disabled": "Disabled", "Enabled": "Enabled"},
+        )
+
+        stop_on_no_connect = cls._args_schema.stop_on_no_connect
+        stop_on_no_connect.grace_period_minutes = AAZIntArg(
+            options=["grace-period-minutes"],
+            help="The specified time in minutes to wait before stopping a Dev Box if no connection is made.",
+        )
+        stop_on_no_connect.status = AAZStrArg(
+            options=["status"],
+            help="Enables the feature to stop a started Dev Box when it has not been connected to, once the grace period has lapsed.",
             enum={"Disabled": "Disabled", "Enabled": "Enabled"},
         )
         return cls._args_schema
@@ -250,7 +321,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-05-01-preview",
+                    "api-version", "2024-10-01-preview",
                     required=True,
                 ),
             }
@@ -281,7 +352,9 @@ class Create(AAZCommand):
 
             properties = _builder.get(".properties")
             if properties is not None:
+                properties.set_prop("devBoxDefinition", AAZObjectType)
                 properties.set_prop("devBoxDefinitionName", AAZStrType, ".devbox_definition_name", typ_kwargs={"flags": {"required": True}})
+                properties.set_prop("devBoxDefinitionType", AAZStrType, ".devbox_definition_type")
                 properties.set_prop("displayName", AAZStrType, ".display_name")
                 properties.set_prop("licenseType", AAZStrType, ".license_type", typ_kwargs={"flags": {"required": True}})
                 properties.set_prop("localAdministrator", AAZStrType, ".local_administrator", typ_kwargs={"flags": {"required": True}})
@@ -289,7 +362,25 @@ class Create(AAZCommand):
                 properties.set_prop("networkConnectionName", AAZStrType, ".network_connection_name", typ_kwargs={"flags": {"required": True}})
                 properties.set_prop("singleSignOnStatus", AAZStrType, ".single_sign_on_status")
                 properties.set_prop("stopOnDisconnect", AAZObjectType, ".stop_on_disconnect")
+                properties.set_prop("stopOnNoConnect", AAZObjectType, ".stop_on_no_connect")
                 properties.set_prop("virtualNetworkType", AAZStrType, ".virtual_network_type")
+
+            dev_box_definition = _builder.get(".properties.devBoxDefinition")
+            if dev_box_definition is not None:
+                dev_box_definition.set_prop("imageReference", AAZObjectType, ".devbox_definition_image_reference")
+                dev_box_definition.set_prop("sku", AAZObjectType, ".devbox_definition_sku")
+
+            image_reference = _builder.get(".properties.devBoxDefinition.imageReference")
+            if image_reference is not None:
+                image_reference.set_prop("id", AAZStrType, ".id")
+
+            sku = _builder.get(".properties.devBoxDefinition.sku")
+            if sku is not None:
+                sku.set_prop("capacity", AAZIntType, ".capacity")
+                sku.set_prop("family", AAZStrType, ".family")
+                sku.set_prop("name", AAZStrType, ".name", typ_kwargs={"flags": {"required": True}})
+                sku.set_prop("size", AAZStrType, ".size")
+                sku.set_prop("tier", AAZStrType, ".tier")
 
             managed_virtual_network_regions = _builder.get(".properties.managedVirtualNetworkRegions")
             if managed_virtual_network_regions is not None:
@@ -299,6 +390,11 @@ class Create(AAZCommand):
             if stop_on_disconnect is not None:
                 stop_on_disconnect.set_prop("gracePeriodMinutes", AAZIntType, ".grace_period_minutes")
                 stop_on_disconnect.set_prop("status", AAZStrType, ".status")
+
+            stop_on_no_connect = _builder.get(".properties.stopOnNoConnect")
+            if stop_on_no_connect is not None:
+                stop_on_no_connect.set_prop("gracePeriodMinutes", AAZIntType, ".grace_period_minutes")
+                stop_on_no_connect.set_prop("status", AAZStrType, ".status")
 
             tags = _builder.get(".tags")
             if tags is not None:
@@ -350,9 +446,15 @@ class Create(AAZCommand):
                 serialized_name="devBoxCount",
                 flags={"read_only": True},
             )
+            properties.dev_box_definition = AAZObjectType(
+                serialized_name="devBoxDefinition",
+            )
             properties.dev_box_definition_name = AAZStrType(
                 serialized_name="devBoxDefinitionName",
                 flags={"required": True},
+            )
+            properties.dev_box_definition_type = AAZStrType(
+                serialized_name="devBoxDefinitionType",
             )
             properties.display_name = AAZStrType(
                 serialized_name="displayName",
@@ -389,9 +491,32 @@ class Create(AAZCommand):
             properties.stop_on_disconnect = AAZObjectType(
                 serialized_name="stopOnDisconnect",
             )
+            properties.stop_on_no_connect = AAZObjectType(
+                serialized_name="stopOnNoConnect",
+            )
             properties.virtual_network_type = AAZStrType(
                 serialized_name="virtualNetworkType",
             )
+
+            dev_box_definition = cls._schema_on_200_201.properties.dev_box_definition
+            dev_box_definition.active_image_reference = AAZObjectType(
+                serialized_name="activeImageReference",
+            )
+            _CreateHelper._build_schema_image_reference_read(dev_box_definition.active_image_reference)
+            dev_box_definition.image_reference = AAZObjectType(
+                serialized_name="imageReference",
+            )
+            _CreateHelper._build_schema_image_reference_read(dev_box_definition.image_reference)
+            dev_box_definition.sku = AAZObjectType()
+
+            sku = cls._schema_on_200_201.properties.dev_box_definition.sku
+            sku.capacity = AAZIntType()
+            sku.family = AAZStrType()
+            sku.name = AAZStrType(
+                flags={"required": True},
+            )
+            sku.size = AAZStrType()
+            sku.tier = AAZStrType()
 
             health_status_details = cls._schema_on_200_201.properties.health_status_details
             health_status_details.Element = AAZObjectType()
@@ -412,6 +537,12 @@ class Create(AAZCommand):
                 serialized_name="gracePeriodMinutes",
             )
             stop_on_disconnect.status = AAZStrType()
+
+            stop_on_no_connect = cls._schema_on_200_201.properties.stop_on_no_connect
+            stop_on_no_connect.grace_period_minutes = AAZIntType(
+                serialized_name="gracePeriodMinutes",
+            )
+            stop_on_no_connect.status = AAZStrType()
 
             system_data = cls._schema_on_200_201.system_data
             system_data.created_at = AAZStrType(
@@ -441,6 +572,27 @@ class Create(AAZCommand):
 
 class _CreateHelper:
     """Helper class for Create"""
+
+    _schema_image_reference_read = None
+
+    @classmethod
+    def _build_schema_image_reference_read(cls, _schema):
+        if cls._schema_image_reference_read is not None:
+            _schema.exact_version = cls._schema_image_reference_read.exact_version
+            _schema.id = cls._schema_image_reference_read.id
+            return
+
+        cls._schema_image_reference_read = _schema_image_reference_read = AAZObjectType()
+
+        image_reference_read = _schema_image_reference_read
+        image_reference_read.exact_version = AAZStrType(
+            serialized_name="exactVersion",
+            flags={"read_only": True},
+        )
+        image_reference_read.id = AAZStrType()
+
+        _schema.exact_version = cls._schema_image_reference_read.exact_version
+        _schema.id = cls._schema_image_reference_read.id
 
 
 __all__ = ["Create"]