+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/pool/_wait.py
@@ -20,7 +20,7 @@ class Wait(AAZWaitCommand):
 
     _aaz_info = {
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/pools/{}", "2024-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/pools/{}", "2024-10-01-preview"],
         ]
     }
 
@@ -136,7 +136,7 @@ class Wait(AAZWaitCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-05-01-preview",
+                    "api-version", "2024-10-01-preview",
                     required=True,
                 ),
             }
@@ -195,9 +195,15 @@ class Wait(AAZWaitCommand):
                 serialized_name="devBoxCount",
                 flags={"read_only": True},
             )
+            properties.dev_box_definition = AAZObjectType(
+                serialized_name="devBoxDefinition",
+            )
             properties.dev_box_definition_name = AAZStrType(
                 serialized_name="devBoxDefinitionName",
                 flags={"required": True},
+            )
+            properties.dev_box_definition_type = AAZStrType(
+                serialized_name="devBoxDefinitionType",
             )
             properties.display_name = AAZStrType(
                 serialized_name="displayName",
@@ -234,9 +240,32 @@ class Wait(AAZWaitCommand):
             properties.stop_on_disconnect = AAZObjectType(
                 serialized_name="stopOnDisconnect",
             )
+            properties.stop_on_no_connect = AAZObjectType(
+                serialized_name="stopOnNoConnect",
+            )
             properties.virtual_network_type = AAZStrType(
                 serialized_name="virtualNetworkType",
             )
+
+            dev_box_definition = cls._schema_on_200.properties.dev_box_definition
+            dev_box_definition.active_image_reference = AAZObjectType(
+                serialized_name="activeImageReference",
+            )
+            _WaitHelper._build_schema_image_reference_read(dev_box_definition.active_image_reference)
+            dev_box_definition.image_reference = AAZObjectType(
+                serialized_name="imageReference",
+            )
+            _WaitHelper._build_schema_image_reference_read(dev_box_definition.image_reference)
+            dev_box_definition.sku = AAZObjectType()
+
+            sku = cls._schema_on_200.properties.dev_box_definition.sku
+            sku.capacity = AAZIntType()
+            sku.family = AAZStrType()
+            sku.name = AAZStrType(
+                flags={"required": True},
+            )
+            sku.size = AAZStrType()
+            sku.tier = AAZStrType()
 
             health_status_details = cls._schema_on_200.properties.health_status_details
             health_status_details.Element = AAZObjectType()
@@ -257,6 +286,12 @@ class Wait(AAZWaitCommand):
                 serialized_name="gracePeriodMinutes",
             )
             stop_on_disconnect.status = AAZStrType()
+
+            stop_on_no_connect = cls._schema_on_200.properties.stop_on_no_connect
+            stop_on_no_connect.grace_period_minutes = AAZIntType(
+                serialized_name="gracePeriodMinutes",
+            )
+            stop_on_no_connect.status = AAZStrType()
 
             system_data = cls._schema_on_200.system_data
             system_data.created_at = AAZStrType(
@@ -286,6 +321,27 @@ class Wait(AAZWaitCommand):
 
 class _WaitHelper:
     """Helper class for Wait"""
+
+    _schema_image_reference_read = None
+
+    @classmethod
+    def _build_schema_image_reference_read(cls, _schema):
+        if cls._schema_image_reference_read is not None:
+            _schema.exact_version = cls._schema_image_reference_read.exact_version
+            _schema.id = cls._schema_image_reference_read.id
+            return
+
+        cls._schema_image_reference_read = _schema_image_reference_read = AAZObjectType()
+
+        image_reference_read = _schema_image_reference_read
+        image_reference_read.exact_version = AAZStrType(
+            serialized_name="exactVersion",
+            flags={"read_only": True},
+        )
+        image_reference_read.id = AAZStrType()
+
+        _schema.exact_version = cls._schema_image_reference_read.exact_version
+        _schema.id = cls._schema_image_reference_read.id
 
 
 __all__ = ["Wait"]