+++ src/network-manager/azext_network_manager/aaz/latest/network/manager/ipam_pool/static_cidr/__cmd_group.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command_group(
     "network manager ipam-pool static-cidr",
-    is_preview=True,
 )
 class __CMDGroup(AAZCommandGroup):
     """Manage static cidr with network.